chatbot/pyproject.toml
[project]
name = "todo-chatbot"
version = "0.1.0"
description = "Phase III: AI-Powered Todo Chatbot with OpenAgents SDK and MCP"
authors = [{ name = "Sharmeen Asif" }]
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "anthropic>=0.40.0",
    "httpx>=0.27.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
    "mcp>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pyright>=1.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pyright]
pythonVersion = "3.13"
typeCheckingMode = "strict"

[project.scripts]
chatbot = "app.main:main"
